# webapp-1/backend/api/swagger.yaml
swagger: "2.0"
info:
  title: Dollbuilder API
  description: RESTful API for the Music Box Doll Builder backend
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: localhost:8080
basePath: /api/v1/
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token. Format: Bearer {token}"
security:
  - jwt: []
paths:
...
  /entitlements:
    get:
      summary: Current user's export entitlement
      operationId: getEntitlement
      tags: [entitlements]
      security: []
      responses:
        200: { description: Entitlement state, schema: { $ref: "#/definitions/Entitlement" } }
        401: { description: Not authenticated, schema: { $ref: "#/definitions/Error" } }
        500: { description: Server error, schema: { $ref: "#/definitions/Error" } }
  /entitlements/mock-checkout:
    post:
      summary: Grant entitlement via the mock path (dev/test only)
      operationId: mockCheckout
      tags: [entitlements]
      security: []
      responses:
        200: { description: Entitlement granted, schema: { $ref: "#/definitions/Entitlement" } }
        401: { description: Not authenticated, schema: { $ref: "#/definitions/Error" } }
        403: { description: Mock checkout disabled, schema: { $ref: "#/definitions/Error" } }
        500: { description: Server error, schema: { $ref: "#/definitions/Error" } }
  /checkout/session:
    post:
      summary: Create a Stripe Checkout session
      operationId: createCheckoutSession
      tags: [checkout]
      security: []
      responses:
        200: { description: Checkout URL, schema: { $ref: "#/definitions/CheckoutSession" } }
        401: { description: Not authenticated, schema: { $ref: "#/definitions/Error" } }
        500: { description: Server error, schema: { $ref: "#/definitions/Error" } }
        503: { description: Checkout not configured, schema: { $ref: "#/definitions/Error" } }
definitions:
  Entitlement:
    type: object
    required: [entitled]
    properties:
      entitled: { type: boolean }
      source: { type: string }
  CheckoutSession:
    type: object
    required: [checkoutUrl]
    properties:
      checkoutUrl: { type: string }
...
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int64
        description: HTTP status code
        x-example: 404
      message:
        type: string
        description: Human-readable error message
        x-example: "Resource not found"
      details:
        type: object
        description: Additional error details (optional)
        additionalProperties: true
        x-example:
          field: "email"
          reason: "invalid format"
...